requirements.txt
streamlit
python-dotenv
google-generativeai
google-cloud-documentai
extract-thinker
Pillow
pydantic
pdf2image
PyMuPDF
